pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset-harvester"
version = "1.0.0"
requires-python = ">=3.10"
description = "Image-to-3D pipeline for autonomous vehicle simulation assets"
license = {text = "Apache-2.0"}
readme = "README.md"
dependencies = [
    "numpy>=1.24.0",
    "Pillow>=12.1.1",
    "PyYAML",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
ncore-parser = [
    "click>=8.2.0",
    "pydantic>=2.0.0",
    "opencv-python>=4.8.0",
    "torch==2.10.0",
    "transformers==4.48.3",
    "timm>=0.9.0",
    "einops>=0.7.0",
    "open-clip-torch",
    "scipy>=1.10.0",
    "wandb>=0.16.0",
    "nvidia-ncore",
]
multiview_diffusion = [
    "torch==2.10.0",
    "torchvision",
    "diffusers>=0.33.0,<0.37.0",
    "transformers==4.48.3",
    "safetensors",
    "imageio>=2.31.0",
    "imageio-ffmpeg>=0.4.0",
    "accelerate>=0.25.0",
    "huggingface_hub>=0.20.0",
    "einops>=0.7.0",
    "webdataset>=0.2.48",
    "omegaconf>=2.3.0",
    "scipy>=1.10.0",
    "opencv-python",
    "timm",
    "open-clip-torch",
    "xformers",
    "packaging",
    "torchmetrics",
]
tokengs = [
    "torch==2.10.0",
    "torchvision",
    "safetensors",
    "imageio>=2.31.0",
    "accelerate>=0.25.0",
    "kiui",
    "roma",
    "lpips",
    "scikit-image",
    "einops>=0.7.0",
    "tyro",
    "packaging",
    "tensorboard",
    "decord",
    "matplotlib",
    "wandb==0.17.5",
    "scipy>=1.10.0",
    "opencv-python",
    "boto3",
]
camera-estimator = [
    "torch==2.10.0",
    "torchvision",
    "transformers==4.48.3",
    "safetensors",
    "omegaconf>=2.3.0",
]

[project.scripts]
ncore-parser = "asset_harvester.ncore_parser.__main__:main"
tokengs-train = "asset_harvester.tokengs.main:main"

[tool.setuptools.packages.find]
include = ["asset_harvester*"]

[tool.ruff]
line-length = 120
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "*.egg-info",
    ".venv",
    "venv",
]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = [
    "E501",
    "E741",
]

[tool.ruff.lint.isort]
known-first-party = ["asset_harvester"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
